fix: sign in button is being disabled whilke we wait for response from backend, on the navbar, login in button is changing to log out when user is signed in
