+++ src/Components/Logout/Logout.module.css
@@ -1,0 +1,6 @@
+.button {
+  background-color: inherit;
+  border: none;
+  color: #fff;
+  font-weight: 600;
+}
