+++ src/pages/Dashboard/Dashboard.jsx
@@ -2,30 +2,14 @@ import React, { useState } from "react";
 import AppLayout from "../../layout/AppLayout";
 import "./Dashboard.css";
 
-import toast from "react-hot-toast";
-
 import { Form, Input } from "antd";
 
 import userImg from "./assets/userImg.png";
-import { useNavigate } from "react-router-dom";
 
 const Dashboard = () => {
   const [isNavExpanded, setIsNavExpanded] = useState(true);
-  const navigate = useNavigate();
   const onFinish = (value) => {
     console.log(value);
-  };
-  const logout = () => {
-    sessionStorage.removeItem("token");
-    navigate("/login");
-    toast.success(`Log out succesful!`, {
-      duration: 4000,
-      position: "top-center",
-
-      // Styling
-      style: { fontSize: "13px" },
-      className: "",
-    });
   };
 
   return (
@@ -87,7 +71,6 @@ const Dashboard = () => {
               <li>
                 <a href="#Part1">Media</a>
               </li>
-              <button onClick={logout}>Log out</button>
             </ul>
           </div>
